requirements.txt
certifi==2020.6.20
Django==4.0.6
djangorestframework==3.13.1
pytz==2022.1
requests==2.25.1
psycopg2==2.9.2
sympy==1.9
